pyproject.toml
[tool.poetry]
name = "kahvikamera-bot"
version = "0.1.0"
description = "Luo kahvikamera botin, joka nyt tekee vähän kaikenlaista"
authors = ["Aleksi <ale.kettunen@gmail.com>"]
license = "Ei"
readme = "README.md"
packages = [{include = "kahvikamera_bot"}]

[tool.poetry.scripts]
kahvikamera-bot = "kahvikamera_bot.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
requests = "^2.28.1"
opencv-python = "^4.6.0.66"
python-telegram-bot = "^13.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
